+++ cmd/common_test.go
@@ -27,7 +27,7 @@ func TestConvDate(t *testing.T) {
 	})
 }
 
-func TestOverThreshold(t *testing.T) {
+func TestOverIntThreshold(t *testing.T) {
 	var num1 int64
 	var num2 int64
 	var alarm bool
@@ -35,7 +35,7 @@ func TestOverThreshold(t *testing.T) {
 	Convey("When the threshold is exceeded", t, func() {
 		num1 = 20
 		num2 = 2
-		alarm = overThreshold(num1, num2)
+		alarm = overIntThreshold(num1, num2)
 
 		Convey("20 should be greater than 2 and the alarm should be true", func() {
 			So(alarm, ShouldBeTrue)
@@ -45,9 +45,35 @@ func TestOverThreshold(t *testing.T) {
 	Convey("When the threshold is not exceeded", t, func() {
 		num1 = 2
 		num2 = 20
-		alarm = overThreshold(num1, num2)
+		alarm = overIntThreshold(num1, num2)
 
 		Convey("2 should be less than 20 and the alarm should not be true", func() {
+			So(alarm, ShouldBeFalse)
+		})
+	})
+}
+
+func TestOverFloatThreshold(t *testing.T) {
+	var num1 float64
+	var num2 float64
+	var alarm bool
+
+	Convey("When the threshold is exceeded", t, func() {
+		num1 = 20.0
+		num2 = 2.0
+		alarm = overFloatThreshold(num1, num2)
+
+		Convey("20.0 should be greater than 2.0 and the alarm should be true", func() {
+			So(alarm, ShouldBeTrue)
+		})
+	})
+
+	Convey("When the threshold is not exceeded", t, func() {
+		num1 = 2.0
+		num2 = 20.0
+		alarm = overFloatThreshold(num1, num2)
+
+		Convey("2.0 should be less than 20.0 and the alarm should not be true", func() {
 			So(alarm, ShouldBeFalse)
 		})
 	})
@@ -142,8 +168,6 @@ func TestCreateMap(t *testing.T) {
 		})
 	})
 }
-
-// 1458528701 1458528828
 
 func TestTimeDeviation(t *testing.T) {
 	Convey("When testing the difference of two times", t, func() {